pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "traverse-backend-wa"
version = "0.0.0"
description = "fastapi service"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.110"
uvicorn = {extras = ["standard"], version = ">=0.29"}
pydantic = {extras = ["email"], version = "^2.12.0"}
python-dotenv = ">=1.0.1"
aisuite = {extras = ["openai"], version = "^0.1.11"}
docstring-parser = ">=0.15"
google-generativeai = ">=0.8.0"
vertexai = "^1.71.1"
httpx = ">=0.27"
requests = "^2.31.0"
pymongo = "^4.15.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
playwright = "^1.55.0"
resend = "^2.5.0"
scikit-learn = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
httpx = ">=0.27"
ruff = ">=0.5"
black = ">=24.3.0"
mypy = ">=1.10"

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = ["B","C4","E","F","I","N","PL","PT","RUF","SIM","TID","UP","W"]
ignore = ["E203","E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false


